requirements.txt
# FastAPI and Server
fastapi==0.109.2
uvicorn==0.27.1
gunicorn==21.2.0
python-multipart==0.0.6

# Authentication and Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Database and ORM
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
pymysql==1.1.0

# API Documentation
pydantic==2.6.1
pydantic-settings==2.1.0

# LangChain and AI
langchain==0.1.9
langchain-openai==0.0.5
openai==1.12.0

# SQL Parsing and Validation
sqlparse==0.4.4
sql-metadata==2.10.0

# Rate Limiting and Caching
redis==5.0.1
limits==3.9.0

# Payment Processing
stripe==7.11.0

# Testing
pytest==8.0.0
pytest-asyncio==0.23.5
httpx==0.26.0
pytest-cov==4.1.0

# Monitoring and Logging
prometheus-client==0.19.0
python-json-logger==2.0.7

# Development Tools
black==24.1.1
isort==5.13.2
flake8==7.0.0

# CORS
fastapi-cors==0.0.6

# Task Queue
celery==5.3.6
flower==2.0.1

# Email
fastapi-mail==1.4.1

# AWS Integration (for deployment)
boto3==1.34.29

